[LayoutNG] Fix OOF not painted in "tall" multicolumns
In some cases as the one described in https://crbug.com/1151383,
non-fragmented OOF are not painted. I'm not an expert in the paint area,
but it seems like we are not always able to detect whether we are in a
fragmentation context or not.

I decided to slightly update our implementation to ensure we're always
going to know whether we are in a fragmentation context or not. Put
simply, I added an |is_in_fragmentation_context| attribute to
PaintLayerRareData.

Bug: 1151383
Change-Id: Ib85440577442782a1f09f704d2224214d4822232